+++ apistar/main.py
@@ -7,8 +7,7 @@ import jinja2
 import apistar
 from apistar import codecs
 from apistar.exceptions import ParseError, ValidationError
-from apistar.formatters.json_errors import json_errors
-from apistar.formatters.yaml_errors import yaml_errors
+from apistar.formatters import get_errors
 
 
 def static_url(filename):
@@ -29,34 +28,33 @@ def validate(schema):
 
     codec = codecs.OpenAPICodec()
     content = schema.read()
-    is_json = content.decode().strip() and content.decode().strip()[0] in '{['
+    if content.decode().strip() and content.decode().strip()[0] in '{[':
+        content_type = 'json'
+    else:
+        content_type = 'yaml'
+
     try:
         codec.decode(content)
-    except ParseError as exc:
-        if is_json:
-            click.echo(click.style('✘', fg='red') + ' Invalid JSON.')
-        else:
-            click.echo(click.style('✘', fg='red') + ' Invalid YAML.')
-        click.echo()
-        if exc.pos is None:
-            click.echo(exc)
-        else:
-            valid_lines = content.splitlines()[:exc.lineno-1]
-            invalid_line = content.splitlines()[exc.lineno-1]
-            remaining_lines = content.splitlines()[exc.lineno:]
+    except (ParseError, ValidationError) as exc:
+        errors = get_errors(content, exc, content_type)
+        lines = content.splitlines()
+        for error in reversed(errors):
+            error_str = ' ' * (error.start.column - 1)
+            error_str += '^ '
+            error_str += error.message
+            error_str = click.style(error_str, fg='red')
+            lines.insert(error.start.row, error_str)
+        for line in lines:
+            click.echo(line)
 
-            click.echo(b'\n'.join(valid_lines))
-            click.echo(invalid_line)
-            click.echo(' ' * (exc.colno - 1), nl=False)
-            click.echo(click.style('^ ' + exc.short_message, fg='red'))
-            click.echo(b'\n'.join(remaining_lines))
-    except ValidationError as exc:
-        click.echo(click.style('✘', fg='red') + ' Invalid OpenAPI 3 document.')
         click.echo()
-        if is_json:
-            click.echo(json_errors(exc.value, exc.detail))
+        if isinstance(exc, ParseError) and content_type == 'json':
+            click.echo(click.style('✘', fg='red') + ' Invalid JSON.')
+        elif isinstance(exc, ParseError) and content_type == 'yaml':
+            click.echo(click.style('✘', fg='red') + ' Invalid YAML.')
         else:
-            click.echo(yaml_errors(exc.value, exc.detail))
+            click.echo(click.style('✘', fg='red') + ' Invalid OpenAPI 3 document.')
+
     else:
         click.echo(click.style('✓', fg='green') + ' Valid OpenAPI 3 document.')