+++ docs/topics/release-notes.md
@@ -2,6 +2,14 @@
 
 ## Version 0.5.x
 
+### 0.5.16
+
+* Improved error formatting.
+
+### 0.5.15
+
+* Fix 'responses' and 'parameters' in OpenAPI spec.
+
 ### 0.5.14
 
 * Added 'headers' to OpenAPI spec.